+++ Journal.lua
@@ -231,23 +231,23 @@ function journal:ADDON_LOADED(addonName)
 			MountJournal_UpdateMountList()
 			self:updateMapSettings()
 			mounts:setMountsList()
-			self.existingsLists:refresh()
+			self.existingLists:refresh()
 
 			self.mountListUpdateAnim:Stop()
 			self.mountListUpdateAnim:Play()
 		end)
 		UIDropDownMenu_Initialize(mapSettings.listFromMap.optionsMenu, self.listFromMapInit, "MENU")
 
-		-- EXISTINGS LISTS TOGGLE
-		mapSettings.existingsListsToggle:HookScript("OnClick", function(btn)
-			self.existingsLists:SetShown(btn:GetChecked())
+		-- EXISTING LISTS TOGGLE
+		mapSettings.existingListsToggle:HookScript("OnClick", function(btn)
+			self.existingLists:SetShown(btn:GetChecked())
 		end)
 
-		-- EXISTINGS LISTS
-		local existingsLists = CreateFrame("FRAME", nil, mapSettings, "MJExistingsListsPanelTemplate")
-		self.existingsLists = existingsLists
-		existingsLists:SetPoint("TOPLEFT", MountJournal, "TOPRIGHT")
-		existingsLists:SetPoint("BOTTOMLEFT", MountJournal, "BOTTOMRIGHT")
+		-- EXISTING LISTS
+		local existingLists = CreateFrame("FRAME", nil, mapSettings, "MJExistingListsPanelTemplate")
+		self.existingLists = existingLists
+		existingLists:SetPoint("TOPLEFT", MountJournal, "TOPRIGHT")
+		existingLists:SetPoint("BOTTOMLEFT", MountJournal, "BOTTOMRIGHT")
 
 		--MOUNTJOURNAL ONSHOW
 		MountJournal:HookScript("OnShow", function()
@@ -295,7 +295,7 @@ function journal:ADDON_LOADED(addonName)
 			self:updateMountsList()
 			MountJournal_UpdateMountList()
 			self:updateMapSettings()
-			self.existingsLists:refresh()
+			self.existingLists:refresh()
 
 			self.mountListUpdateAnim:Stop()
 			self.mountListUpdateAnim:Play()
@@ -804,15 +804,13 @@ end
 function journal:grid3UpdateMountList()
 	local scrollFrame = self.scrollFrame
 	local offset = HybridScrollFrame_GetOffset(scrollFrame)
-	local numButtons = #scrollFrame.buttons
 	local numDisplayedMounts = C_MountJournal.GetNumDisplayedMounts()
+	local selectedSpellID = MountJournal.selectedSpellID
 
 	for i, btn in ipairs(scrollFrame.buttons) do
 		for j = 1, 3 do
 			local index = (offset + i - 1) * 3 + j
 			local btnGrid = btn.grid3list["mount"..j]
-
-			btnGrid.icon:SetVertexColor(1, 1, 1)
 
 			if index <= numDisplayedMounts then
 				local creatureName, spellID, icon, active, isUsable, sourceType, isFavorite, isFactionSpecific, faction, isFiltered, isCollected, mountID = C_MountJournal.GetDisplayedMountInfo(index)
@@ -821,8 +819,9 @@ function journal:grid3UpdateMountList()
 				btnGrid.index = index
 				btnGrid.spellID = spellID
 				btnGrid.active = active
-				btnGrid.selected = MountJournal.selectedSpellID == spellID
+				btnGrid.selected = selectedSpellID == spellID
 				btnGrid.icon:SetTexture(needsFanfare and COLLECTIONS_FANFARE_ICON or icon)
+				btnGrid.icon:SetVertexColor(1, 1, 1)
 				btnGrid.favorite:SetShown(isFavorite)
 				btnGrid.DragButton:Enable()
 				btnGrid.DragButton.selectedTexture:SetShown(btnGrid.selected)
@@ -874,18 +873,16 @@ function journal:setEditMountsList()
 			ground = self.db.ground,
 			swimming = self.db.swimming,
 		}
-		self.list = self.currentList
 		self.listMapID = nil
+		self.list = self.currentList
 	else
-		local function getRelationMountList(mapID)
-			local list = self.db.zoneMounts[mapID]
-			if list and list.listFromID then
-				return getRelationMountList(list.listFromID)
-			end
-			return list, mapID
-		end
 		self.currentList = self.db.zoneMounts[mapID]
-		self.list, self.listMapID = getRelationMountList(mapID)
+		self.listMapID = mapID
+		self.list = self.currentList
+		while self.list and self.list.listFromID do
+			self.listMapID = self.list.listFromID
+			self.list = self.db.zoneMounts[self.listMapID]
+		end
 	end
 end
 
@@ -1010,7 +1007,7 @@ function journal:mountToggle(btn)
 	end
 
 	mounts:setMountsList()
-	self.existingsLists:refresh()
+	self.existingLists:refresh()
 end
 
 
@@ -1026,7 +1023,7 @@ function journal:setFlag(flag, enable)
 	end
 
 	mounts:setMountsList()
-	self.existingsLists:refresh()
+	self.existingLists:refresh()
 end
 
 
@@ -1095,7 +1092,7 @@ function journal:listFromMapInit(level)
 			MountJournal_UpdateMountList()
 			journal:updateMapSettings()
 			mounts:setMountsList()
-			journal.existingsLists:refresh()
+			journal.existingLists:refresh()
 			CloseDropDownMenus()
 
 			journal.mountListUpdateAnim:Stop()
@@ -1678,7 +1675,7 @@ end
 
 
 function journal:updateMountsList()
-	local types, selected, factions, pet, expansions, list, tags, GetDisplayedMountInfo, GetMountInfoExtraByID = mounts.filters.types, mounts.filters.selected, mounts.filters.factions, mounts.filters.pet, mounts.filters.expansions, self.list, self.tags, self.func.GetDisplayedMountInfo, C_MountJournal.GetMountInfoExtraByID
+	local types, selected, factions, pet, expansions, mountTypes, list, tags, inTable, GetDisplayedMountInfo, GetMountInfoExtraByID = mounts.filters.types, mounts.filters.selected, mounts.filters.factions, mounts.filters.pet, mounts.filters.expansions, self.mountTypes, self.list, self.tags, util.inTable, self.func.GetDisplayedMountInfo, C_MountJournal.GetMountInfoExtraByID
 	wipe(self.displayedMounts)
 
 	for i = 1, self.func.GetNumDisplayedMounts() do
@@ -1688,17 +1685,17 @@ function journal:updateMountsList()
 		mountFaction = mountFaction or 2
 
 		-- TYPE
-		if types[self.mountTypes[mountType]]
+		if types[mountTypes[mountType]]
 		-- FACTION
 		and factions[mountFaction + 1]
 		-- SELECTED
 		and (not selected[1] and not selected[2] and not selected[3]
 			-- FLY
-			or selected[1] and list and util.inTable(list.fly, mountID)
+			or selected[1] and list and inTable(list.fly, mountID)
 			-- GROUND
-			or selected[2] and list and util.inTable(list.ground, mountID)
+			or selected[2] and list and inTable(list.ground, mountID)
 			-- SWIMMING
-			or selected[3] and list and util.inTable(list.swimming, mountID))
+			or selected[3] and list and inTable(list.swimming, mountID))
 		-- PET
 		and pet[petID and (type(petID) == "number" and petID or 3) or 4]
 		-- EXPANSIONS