+++ _test.lua
@@ -35,6 +35,7 @@ end)
 function test:PLAYER_ENTERING_WORLD()
 	self:UnregisterEvent("PLAYER_ENTERING_WORLD")
 	fprint(GetBuildInfo())
+
 	-- JOURNAL OPEN
 	-- if not IsAddOnLoaded("Blizzard_Collections") then
 	-- 	LoadAddOn("Blizzard_Collections")
@@ -177,5 +178,8 @@ end
 -- C_PetJournal.SummonPetByGUID(PetJournalPetCard.petID)
 -- C_PetJournal.GetSummonedPetGUID()
 
--- debugprofilestart() - starts a timer for profiling during debugging.
--- debugprofilestop() - return the time in milliseconds since the last call to debugprofilestart()
+-- DEBUG IN MILLISECONDS
+-- debugprofilestart()
+-- for i = 1, 10000 do
+-- end
+-- fprint(debugprofilestop())