+++ mixins/MacroMixin.lua
@@ -45,6 +45,7 @@ function macroFrame:refresh()
 	self:setCombatMacro()
 end
 
+
 function macroFrame:addLine(text, line)
 	if type(text) == "string" and text:len() > 0 then
 		return strjoin("\n", text, line)
@@ -181,8 +182,8 @@ do
 	}
 
 
-	function macroFrame:getMacro(forceModifier)
-		self.mounts:setFlags(forceModifier)
+	function macroFrame:getMacro()
+		self.mounts:setFlags()
 		local macro
 
 		-- DRUID LAST FORM
@@ -191,11 +192,6 @@ do
 		-- 24858 - moonkin form
 		if self.classConfig.useLastDruidForm then
 			local spellID = getFormSpellID()
-
-			if GetShapeshiftFormID() == 3 and (GetUnitSpeed("player") > 0 or IsFalling()) then
-				macro = "/cancelform"
-				return macro or ""
-			end
 
 			if self.classConfig.useDruidFormSpecialization then
 				self.lastDruidFormSpellID = specializationSpellIDs[GetSpecialization()]
@@ -286,13 +282,15 @@ end
 
 
 function MJMacroMixin:onLoad()
+	self.mounts = MountsJournal
 	self:RegisterEvent("PLAYER_REGEN_DISABLED")
 end
 
 
 function MJMacroMixin:preClick()
+	self.mounts.sFlags.forceModifier = self.forceModifier
 	if InCombatLockdown() then return end
-	self:SetAttribute("macrotext", macroFrame:getMacro(self.forceModifier))
+	self:SetAttribute("macrotext", macroFrame:getMacro())
 end