+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* Fixed tag sorting.
+* Added search on existing lists.
+* Minor fixes.